requirements.txt
flask
opencv-python
numpy
pandas
redis
insightface
scikit-learn
python-dotenv
onnxruntime
geopy